# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Finca Farm Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farms/{farmId}/animals:
    get:
      operationId: listAnimals
      tags: [animals]
      summary: List animals for a farm
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: species
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: breed
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of animals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Animal"
    post:
      operationId: createAnimal
      tags: [animals]
      summary: Add an animal to the farm
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnimalRequest"
      responses:
        "201":
          description: Created animal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Animal"
  /farms/{farmId}/animals/{animalId}:
    get:
      operationId: getAnimal
      tags: [animals]
      summary: Get animal details
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: animalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Animal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimalDetail"
    put:
      operationId: updateAnimal
      tags: [animals]
      summary: Update animal
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: animalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnimalRequest"
      responses:
        "200":
          description: Updated animal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Animal"
  /farms/{farmId}/animals/{animalId}/weights:
    get:
      operationId: listWeightRecords
      tags: [weights]
      summary: List weight records for an animal
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: animalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Weight records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightRecord"
    post:
      operationId: createWeightRecord
      tags: [weights]
      summary: Add weight record
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: animalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWeightRecordRequest"
      responses:
        "201":
          description: Created weight record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightRecord"
  /farms/{farmId}/animals/{animalId}/medical:
    get:
      operationId: listMedicalRecords
      tags: [medical]
      summary: List medical records for an animal
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: animalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Medical records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalRecord"
    post:
      operationId: createMedicalRecord
      tags: [medical]
      summary: Add medical record
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: animalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicalRecordRequest"
      responses:
        "201":
          description: Created medical record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
...
  /farms/{farmId}/activity:
    get:
      operationId: listActivity
      tags: [activity]
      summary: List activity log for a farm
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Activity log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEntry"
...
components:
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [owner, worker]
        preferredLanguage:
          type: string
        createdAt:
          type: string
      required:
        - id
...
    Animal:
      type: object
      properties:
        id:
          type: string
        farmId:
          type: string
        customTag:
          type: string
        species:
          type: string
          enum: [cattle, chicken, pig, goat, sheep, horse, other]
        breed:
          type: string
        name:
          type: string
        sex:
          type: string
          enum: [male, female, unknown]
        dateOfBirth:
          type: string
        status:
          type: string
          enum: [active, sold, deceased, transferred]
        photoUrl:
          type: string
        notes:
          type: string
        currentWeight:
          type: number
        currentZoneId:
          type: string
        createdAt:
          type: string
      required:
        - id
        - farmId
        - species
    AnimalDetail:
      allOf:
        - $ref: "#/components/schemas/Animal"
        - type: object
          properties:
            weightRecords:
              type: array
              items:
                $ref: "#/components/schemas/WeightRecord"
            medicalRecords:
              type: array
              items:
                $ref: "#/components/schemas/MedicalRecord"
            offspring:
              type: array
              items:
                $ref: "#/components/schemas/Animal"
            mother:
              $ref: "#/components/schemas/Animal"
            father:
              $ref: "#/components/schemas/Animal"
    CreateAnimalRequest:
      type: object
      properties:
        customTag:
          type: string
        species:
          type: string
          enum: [cattle, chicken, pig, goat, sheep, horse, other]
        breed:
          type: string
        name:
          type: string
        sex:
          type: string
          enum: [male, female, unknown]
        dateOfBirth:
          type: string
        status:
          type: string
          enum: [active, sold, deceased, transferred]
        motherId:
          type: string
        fatherId:
          type: string
        currentZoneId:
          type: string
        notes:
          type: string
      required:
        - species
    UpdateAnimalRequest:
      type: object
      properties:
        customTag:
          type: string
        species:
          type: string
        breed:
          type: string
        name:
          type: string
        sex:
          type: string
        dateOfBirth:
          type: string
        status:
          type: string
        motherId:
          type: string
        fatherId:
          type: string
        currentZoneId:
          type: string
        photoUrl:
          type: string
        notes:
          type: string
    WeightRecord:
      type: object
      properties:
        id:
          type: string
        animalId:
          type: string
        weightKg:
          type: number
        recordedAt:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - animalId
        - weightKg
        - recordedAt
    CreateWeightRecordRequest:
      type: object
      properties:
        weightKg:
          type: number
        recordedAt:
          type: string
        notes:
          type: string
      required:
        - weightKg
    MedicalRecord:
      type: object
      properties:
        id:
          type: string
        animalId:
          type: string
        recordType:
          type: string
          enum: [vaccination, treatment, checkup, surgery, deworming, other]
        title:
          type: string
        description:
          type: string
        vetName:
          type: string
        costCop:
          type: number
        recordDate:
          type: string
        nextDueDate:
          type: string
        createdAt:
          type: string
        animalName:
          type: string
        animalTag:
          type: string
      required:
        - id
        - animalId
        - recordType
        - title
        - recordDate
    CreateMedicalRecordRequest:
      type: object
      properties:
        recordType:
          type: string
          enum: [vaccination, treatment, checkup, surgery, deworming, other]
        title:
          type: string
        description:
          type: string
        vetName:
          type: string
        costCop:
          type: number
        recordDate:
          type: string
        nextDueDate:
          type: string
      required:
        - recordType
        - title
        - recordDate
...
    ActivityEntry:
      type: object
      properties:
        id:
          type: string
        farmId:
          type: string
        userId:
          type: string
        actionType:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        description:
          type: string
        createdAt:
          type: string
        profile:
          $ref: "#/components/schemas/Profile"
      required:
        - id
        - farmId
        - actionType
        - entityType
...