# lib/api-zod/src/generated/types/createAnimalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finca Farm Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAnimalRequestSex } from "./createAnimalRequestSex";
import type { CreateAnimalRequestSpecies } from "./createAnimalRequestSpecies";
import type { CreateAnimalRequestStatus } from "./createAnimalRequestStatus";

export interface CreateAnimalRequest {
  customTag?: string;
  species: CreateAnimalRequestSpecies;
  breed?: string;
  name?: string;
  sex?: CreateAnimalRequestSex;
  dateOfBirth?: string;
  status?: CreateAnimalRequestStatus;
  motherId?: string;
  fatherId?: string;
  currentZoneId?: string;
  notes?: string;
}