# lib/api-zod/src/generated/types/updateAnimalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finca Farm Management API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAnimalRequest {
  customTag?: string;
  species?: string;
  breed?: string;
  name?: string;
  sex?: string;
  dateOfBirth?: string;
  status?: string;
  motherId?: string;
  fatherId?: string;
  currentZoneId?: string;
  photoUrl?: string;
  notes?: string;
}